Improve initialization
