+++ src/NuGet.Services.Revalidate/Program.cs
@@ -10,7 +10,7 @@ namespace NuGet.Services.Revalidate
         static void Main(string[] args)
         {
             var job = new Job();
-            JobRunner.Run(job, args).GetAwaiter().GetResult();
+            JobRunner.RunOnce(job, args).GetAwaiter().GetResult();
         }
     }
 }
